requirements.txt
django
djangorestframework
django-filter
gunicorn
django-heroku
psycopg2-binary